requirements.txt
faiss==1.9.0
lightgbm==4.6.0
matplotlib==3.10.0
matplotlib-venn==1.1.2
numpy
pandas==2.2.3
pillow
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
statannot==0.2.3
statannotations==0.7.2
statsmodels==0.14.4
timm==1.0.12
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
